+++ src/js/textMode/graphic.js
@@ -696,6 +696,10 @@ Graphic.prototype = {
 
   setCurrentFrame: function(frame) {
 
+    if(this.editor.commands) {
+      this.editor.commands.setFrame(frame);
+    }
+
     if(this.currentFrame === frame) {
       return;
     }