+++ src/js/textMode/tileSet/tileSet.js
@@ -3034,36 +3034,29 @@ TileSet.prototype = {
 
 
     var oldValue = this.getPixel(character, x, y);
-
     var pos = x + y * this.charWidth;
-    this.tileData[character].data[0][pos] = set;
-
-//    if(updateCharacter) {
-    this.updateCharacterCurrentData(character);
-
-//    }
-
-
-
-//    this.setImageDataPixel(character, x, y, set);
-
-    // record the action
-    var params = { 
-                   "c": character,
-                   "x": x, "y": y,
-                   "oldValue": oldValue,
-                   "newValue": set
-                  }
-
-    this.editor.history.addAction("setCharPixel", params);
-
-
-    if(updateCharacter) {
-      this.updateCharacter(character);
-    }
-    this.modified();    
-
-    this.customCharacterset = true;
+    var _this = this;
+    return this.editor.commands.executeTilePixelEdit({
+      character: character,
+      x: x,
+      y: y,
+      frame: frame,
+      oldValue: oldValue,
+      newValue: set,
+      apply: function() {
+        _this.tileData[character].data[0][pos] = set;
+        _this.updateCharacterCurrentData(character);
+        _this.customCharacterset = true;
+      },
+      markDirty: function() {
+        _this.modified();
+      },
+      invalidate: function() {
+        if(updateCharacter) {
+          _this.updateCharacter(character);
+        }
+      }
+    });
   },
 
   // call this when a character has been updated